remove zram and move tor to settings
