+++ web/projects/ui/src/app/app/preloader/preloader.component.ts
@@ -40,8 +40,6 @@ const ICONS = [
   'file-tray-stacked-outline',
   'finger-print-outline',
   'flash-outline',
-  'flask-outline',
-  'flash-off-outline',
   'folder-open-outline',
   'globe-outline',
   'grid-outline',
@@ -70,6 +68,7 @@ const ICONS = [
   'receipt-outline',
   'refresh',
   'reload',
+  'reload-circle-outline',
   'remove',
   'remove-circle-outline',
   'remove-outline',
