+++ web/projects/ui/src/app/pages/server-routes/server-routing.module.ts
@@ -80,13 +80,6 @@ const routes: Routes = [
     loadChildren: () =>
       import('./wifi/wifi.module').then(m => m.WifiPageModule),
   },
-  {
-    path: 'experimental-features',
-    loadChildren: () =>
-      import('./experimental-features/experimental-features.module').then(
-        m => m.ExperimentalFeaturesPageModule,
-      ),
-  },
 ]
 
 @NgModule({
