+++ web/projects/ui/src/app/services/api/api.types.ts
@@ -77,11 +77,6 @@ export module RR {
   } // net.tor.reset
   export type ResetTorRes = null
 
-  export type ToggleZramReq = {
-    enable: boolean
-  } // server.experimental.zram
-  export type ToggleZramRes = null
-
   // sessions
 
   export type GetSessionsReq = {} // sessions.list
