+++ web/projects/ui/src/app/services/api/embassy-api.service.ts
@@ -97,8 +97,6 @@ export abstract class ApiService {
 
   abstract resetTor(params: RR.ResetTorReq): Promise<RR.ResetTorRes>
 
-  abstract toggleZram(params: RR.ToggleZramReq): Promise<RR.ToggleZramRes>
-
   // marketplace URLs
 
   abstract marketplaceProxy<T>(
