+++ web/projects/ui/src/app/services/api/embassy-live-api.service.ts
@@ -189,10 +189,6 @@ export class LiveApiService extends ApiService {
     return this.rpcRequest({ method: 'net.tor.reset', params })
   }
 
-  async toggleZram(params: RR.ToggleZramReq): Promise<RR.ToggleZramRes> {
-    return this.rpcRequest({ method: 'server.experimental.zram', params })
-  }
-
   // marketplace URLs
 
   async marketplaceProxy<T>(
