+++ web/projects/ui/src/app/services/api/embassy-mock-api.service.ts
@@ -370,20 +370,6 @@ export class MockApiService extends ApiService {
     return null
   }
 
-  async toggleZram(params: RR.ToggleZramReq): Promise<RR.ToggleZramRes> {
-    await pauseFor(2000)
-    const patch = [
-      {
-        op: PatchOp.REPLACE,
-        path: '/server-info/zram',
-        value: params.enable,
-      },
-    ]
-    this.mockRevision(patch)
-
-    return null
-  }
-
   // marketplace URLs
 
   async marketplaceProxy(
