+++ web/projects/ui/src/app/services/api/mock-patch.ts
@@ -73,7 +73,6 @@ export const mockPatchData: DataModel = {
     pubkey: 'npub1sg6plzptd64u62a878hep2kev88swjh3tw00gjsfl8f237lmu63q0uf63m',
     'ca-fingerprint': 'SHA-256: 63 2B 11 99 44 40 17 DF 37 FC C3 DF 0F 3D 15',
     'ntp-synced': false,
-    zram: false,
     platform: 'x86_64-nonfree',
   },
   'package-data': {
